#275: Image not showing
The logo image for the GeoNetwork site was only created the first time when the database was initialized. 

If the application was updated with a new version (images/logos folder only contains dummy.gif logo then) and preserving the database, the logo was not created again.

Changed to check always in startup if logo for GeoNetwork site exists, if not then creates it

git-svn-id: https://geonetwork.svn.sourceforge.net/svnroot/geonetwork/trunk@6579 ff403467-3f20-0410-99e5-ff581e626a08