[Jeeves] properly take into account HTTP codes
Now that the security-proxy proxifies HTTP responses even if the status
code is different from 200 OK, it is possible from GeoNetwork to send a
default image (if not found), with the 404 status code. But Jeeves
handles HTTP codes only in case of exception / error reportings.

This patch allows to take into account the error code provided by
GeoNetwork into Jeeves. For example if an image is not found by the GN's
resources.get service, it is possible to send to the client a 404
response with an image as part of the body.

This allows to track down broken images in the ComputeDeadLinks service
correctly.

This patch has been tested on
http://sdi.qualitystreetmap.org/geonetwork/ before commiting.