issue #751 - Group parameter
this introduces the ability to select in which group the MD has to be
published (by sending the group_id parameter). If parameter is missing,
we still continue to (try to) publish into the group "1", but the
current user shall have access to the group, or else it will fail.