WxS harvesting - Do not remove MDD across harvests
- The Service metadata generated for the WxS harvested node is still
  deleted / recreated, because it is more simpler than having to sweep
  off child MD references. Note: the uuid is computed from the URL, so
  it is supposed to be stable.
- MDD still present in the catalogue but not in the remote
  getcapabilities response are removed at the end of the process.
- the update process might require a little bit more testing.

Tests: Due to the current state of the code (and the complexity of the
dynamic aspects of a complete testsuite for such usecases), it would
have required a complete rewrite to allow better testing, I could not
manage to make it more testable than the introduced HarvesterTest.java
class. the code has been runtime tested, and should be pretty safe
though.

Portability node: porting the current code into GN3 will require at
least 3 days of work, and might be the occasion to rewrite the harvester
in a more testable fashion. For now, I consider keeping it
"geOrchestra-specific".

Should address the following issues:
https://github.com/camptocamp/georchestra-cigalsace-configuration/issues/340
https://github.com/camptocamp/georchestra-geopicardie-configuration/issues/411