WxS harvesting - use basic-auth when getting MDD
When harvesting, if the metadataUrl for the data metadata is hosted on
the same host as the one for the getcapabilities URL, we can use the
same basic-auth as the one configured for the harvesting endpoint.

A new option is added to the harvesting interface, to enable or disable
this behaviour.

Tests: untested for now