gn2 - dockerization
Contrary to the other webapps in georchestra, we need to use tomcat,
because we can predict where the webapp will be unzipped
(/usr/local/tomcat/webapps/geonetwork/). This is not the case with
Jetty, which unzips into /tmp and uses a timestamp in the generated
directory.

The need to know where the file will be available comes from the fact
that we use an env variable for the jeeves overrides file which is
nested into the webapp (see the added setenv.sh file from this commit
for more details).

Tests: manually on pigma dev.

The same convention applies for the generated image, e.g. you can use
-DdockerImageName=myorg/mygeonetwork2:tag as maven property.