CSW server - Fix proposal for upstream #2069
It seems currently impossible to define a "MUST NOT" match on virtual
csw criteria. this seems related to a bug in the lucene version used
here and on GN upstream, see:

https://github.com/geonetwork/core-geonetwork/issues/2069

This has been reported by INRA to the geOrchestra support (private
issue):

https://jira.camptocamp.com/browse/GEO-483

This fix might not be the most elegant one, but this should do the job:
we now use a customqueryparser along with a custom syntax parser instead
of the buggy one. also the CswCustomSyntaxParser introduced is basically
a copy/paste from the fixed Lucene class StandardSyntaxParser.

Also the StandardQueryParser used before seemed to be called only in
GetRecords related operations, I could not find any references
elsewhere.

Tests: utests added / passed