UploadAndProcess - correctly calculate the resources.get URL
When uploading a file, the update-fixed-info.xsl was responsible of
rewriting the urls so that the link to download the file from GeoNetwork
was correct. But this was breaking the fact that the resource could be
located elsewhere (outside of GeoNetwork services).

The previous commits aimed to remove the URL rewrite, but then after a
successful upload, only the filename was entered in the gmd:URL XML
element. This commit allows to calculate correctly the URL so that the
file is fully qualified.

Tests: utests added, runtime tests on dev.pigma.org.

Related (private) C2C JIRA issue:
https://jira.camptocamp.com/browse/GEO-699