Harvester / Avoid to set empty XSL transformation
Related to https://github.com/geonetwork/core-geonetwork/commit/4193933ea0f19daae75c2050f8bf679d582ab086

In some cases, the element is an object and the empty line was added in such case.

Only setup the element value to the first transformation if it is not set to avoid a reset. In debug mode, it was working but not in prod mode.

Cache the call to get the list of transformation.