Revert "LDAP / Preserve user privileges on sign in."
This reverts commit 7461c029ba40af22c1902d46b144613a347a6ef9.

In geOrchestra, we actually want to trust the LDAP which is managed by
the console webapp, and expect GeoNetwork to apply what is set (from the
Console) in the LDAP, in regards to the users' group ownership and
privileges. As a result, we want to get back to the previous behaviour.