UserSavedSelections / index amount of users that saved a record
Backport of https://github.com/geonetwork/core-geonetwork/pull/6518

A new userSavedCount field is added to the records in the index. It contains
the amount of users that saved the record in their preferred records list.
This also adds indexation of records that are added/removed to saved
selections. Note that records are not indexed when a whole selection is cleared
or deleted.