fix: use application/octet-stream as default mimetype
Using a too clever mime-type guessing could lead to have spring try to
serialize resources using Jackson.

For example, if a JSON file is sent via the AttachmentsApi, then
retrieving it will lead to an error, as the mimetype of the response
will be preemptively set to "application/json", which will trigger
Spring to (try to) use Jackson to serialize the resources, which does
not make sense when the service should just be about dumping the file
into the ServletResponse object.

This patch also drops the `Files.probeContentType()` from the Java API,
which would be also too clever here.

Strategy here is to keep as dumb as possible, to avoid triggering too
much cleverness from the framework.

Tests: runtime tested on dev infrastructure.